+++ .idea/artifacts/desktop_jvm_1_0_SNAPSHOT.xml
@@ -1,8 +1,8 @@
 <component name="ArtifactManager">
   <artifact type="jar" name="desktop-jvm-1.0-SNAPSHOT">
-    <output-path>$PROJECT_DIR$/desktop/build/libs</output-path>
+    <output-path>$PROJECT_DIR$/sample/desktop/build/libs</output-path>
     <root id="archive" name="desktop-jvm-1.0-SNAPSHOT.jar">
-      <element id="module-output" name="compose-richeditor.desktop.jvmMain" />
+      <element id="module-output" name="compose-richeditor.sample.desktop.jvmMain" />
     </root>
   </artifact>
 </component>
