+++ .idea/artifacts/richeditor_desktop_1_0_SNAPSHOT.xml
@@ -1,0 +1,8 @@
+<component name="ArtifactManager">
+  <artifact type="jar" name="richeditor-desktop-1.0-SNAPSHOT">
+    <output-path>$PROJECT_DIR$/richeditor/build/libs</output-path>
+    <root id="archive" name="richeditor-desktop-1.0-SNAPSHOT.jar">
+      <element id="module-output" name="compose-richeditor.richeditor.desktopMain" />
+    </root>
+  </artifact>
+</component>
