+++ sample/desktop/src/jvmMain/kotlin/Main.kt
@@ -1,6 +1,6 @@
 import androidx.compose.ui.window.Window
 import androidx.compose.ui.window.application
-import com.mocoding.richeditor.App
+import com.mocoding.richeditor.sample.common.App
 
 
 fun main() = application {
